fix(config): font_size is of type f32, not u64

+++ src/structs/config.rs
@@ -9,7 +9,7 @@ pub struct ConfigChanged {
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct ConfigChanges {
     pub font_face: Option<String>,
-    pub font_size: Option<u64>,
+    pub font_size: Option<f32>,
     pub line_ending: Option<String>,
     pub plugin_search_path: Option<Vec<String>>,
     pub tab_size: Option<u64>,
